+++ CrossTestDocumenter/CrossTestDocumenter.cs
@@ -12,7 +12,7 @@ namespace CrossTestDocumenter
 
             // Load CrossBreeze CrossTest assembly
             // Loading from the CrossTestdocumenter debug folder otherwise the getCustomAttribute stepDocAttribute cannot be loaded
-            Assembly configAssembly = Assembly.LoadFrom("C:\\GIT\\Repos\\CrossBreeze\\CrossTest\\CrossTest\\CrossTestDocumenter\\bin\\Debug\\CrossBreeze.CrossTest.SpecFlow.dll");
+            Assembly configAssembly = Assembly.LoadFrom("C:\\GIT\\Repos\\CrossBreeze\\CrossTest\\Crosstest_DotNet_SpecFlow\\Lib\\CrossBreeze_CrossTest_SpecFlow_Core\\bin\\Debug\\CrossBreeze.CrossTest.SpecFlow.Core.dll");
 
             //Assembly databaseConfigAssembly = Assembly.LoadFrom("C:\\GIT\\Repos\\CrossBreeze\\CrossTest\\CrossTest\\CrossTestDocumenter\\bin\\Debug\\CrossBreeze.CrossTest.Database.dll");
             //Assembly ssisProcessConfigAssembly = Assembly.LoadFrom("C:\\GIT\\Repos\\CrossBreeze\\CrossTest\\CrossTest\\CrossTestDocumenter\\bin\\Debug\\CrossBreeze.CrossTest.Process.Ssis.dll");
@@ -27,14 +27,16 @@ namespace CrossTestDocumenter
             new CrossTestConfigDocumenter().CreateConfigDocumentation
             (
                 configType,
-                "C:\\GIT\\Repos\\CrossBreeze\\CrossTest\\CrossTest\\Documentation\\generator\\Model\\CrossTestConfig.xml"
+                "C:\\GIT\\Repos\\CrossBreeze\\CrossTest\\CrossTest\\Generation\\Model\\CrossTestConfig.xml"
             );
 
             // Create the Specflow Steps documentation.
-            new CrossTestStepsDocumenter().CreateSpecflowDocumentation(
+            // Replaced by documentation from PDM
+            /*new CrossTestStepsDocumenter().CreateSpecflowDocumentation(
                 configAssembly,
                 "C:\\GIT\\Repos\\CrossBreeze\\CrossTest\\CrossTest\\Documentation\\generator\\Model\\CrossTestSpec.xml"
             );
+            */
 
         }